Don't overzoom when selecting a large POI
Previously we zoomed to the "point" of any POI to a zoom level that
seemed optimized for building-sized points. For large features, like an
entire city, this meant zooming way in to some seemingly random part of
the city.

Now we prefer the bbox whenever available instead of the point.

With this commit, zooming works reasonably for city-sized objects and
the node-based POIs, which have no bbox such as:

e.g. Overcast Coffee Company - "id": "node/4349270303",

But some small venues are defined as an osm way, and thus have a Bbox. With this change
we're now zooming a little too close (IMO). An example of a too-close zoom:

e.g. Victrola Coffee Roasters - "id": "way/206623301"