clearer code
Two-way data bindings make it hard to determine cause and effect. Now we
ensure that our "watch" code is only called in response to changes of
the SearchBox, whereas before it was *also* called during initial load.

I also think that now that POI is being handed in, it can't be
undefined within watch, so no need for that fallback.

Further: was their any reason for the async dispatch?