Start zoomed on last location when possible.
To do this, we store previous map center and zoom in localStorage.

There's some special care taken to not store/fetch too often. Maybe this
is overkill with the debounce, but it seemed wasteful to potentially
trigger i/o as the user pans around or zooms in.

We also recall the users previous zoom - we wouldn't want to zoom in
*further* then they were previously zoomed. That could be particularly
disorienting if the user was at a global scale looking at an ocean or
something.

TBH this risk still exists if the user accidentally zooms way into an
ocean before closing their browser. It'd be nice to have some kind of
magic "is user staring at the ocean" check to short circuit this
behavior, but I _think_, even without that, the risk is worth trying out
the new UI.