use "from" location to determine preferred distance units
There's an unfortunate redundant fetch to the pelias endpoint to
populate the POI. I'd like to address this in future work, probably by
replacing POI with Place, but POI is used all over the place so I'd want
to do it gradually as I have a better understanding of how its used.

In the meanwhile I think this extra request is worth getting the units
right for people.