[www] remove ripple animations
It looks really bad on low-end devices (like my phone).

It's misleading in the steps page, because there is not interaction yet.

It's inconsistent on the mode-picker buttons because it shows when
switching modes within Alternates, but there is no ripple when switching
to the multimodal page (and vice versa).

And, I guess also I'm just asserting my spartan design preferences.