[deploy] inline pelias config
Previously the pelias config lived as an external artifact which was
downloaded on startup.

This had some downsides, primarily that it included hostnames of our
cluster.

It's not a large asset, so we instead bake it into our k8s
deployment-config.

Now we can update our hostnames within the containers and k8s deployment
config and don't have to coordinate those with any uploaded config
artifact.