[www] Layout overhaul to support Desktop
Previously everything was optimized around a mobile layout. Now we have
a layout optimized for desktop too.

I've also made some small improvements for the mobile layout, like we
don't use "rounded" corners for our 100% width cards, which looked
weird.

I've also removed the "bottom card allowance" manual layout accounting.
It was kind of tricky, but primarily: We now use a flexbox layout, and
leverage the `order` attribute to switch between the mobile and desktop
layouts.

And then there is some special handling for the front page, which has a
unique layout.