[deploy] better encapsulate tileserver responsibilities
This is just shuffling configuration around, there's no real change in
behavior, but I hope the following makes things easier to reason about:

All requests to the tileserver are now prefixed with /tileserver.
Routing /data, /styles, and /fonts to the tileserver was a little
unintuitive IMO.

Fonts and sprites moved from the nginx frontend to the tileserver, the
goal was to make the tileserver more "self contained" rather than
scattering details of it's operational requirements across multiple
nodes.

By using the tileserver -u option, we no longer have to template the
HEADWAY_PUBLIC_URL into style.json, so style.json is now a static file.

Fonts and sprites are now baked into the tileserver container, rather
than downloaded at runtime. This is less dynamic, e.g. if someone wanted
to swap out a font or sprite, but I think that's reasonable given that
style.json is already baked into the init-image. The benefits are less
configuration and less happening at application startup.