[build] fix some roads not visible
This is a fixup to 8d15b19

Previously we were templating the style file, but that's no longer
necessarily. However, because it was a template, there were some escaped
characters (${ESC} -> $). I failed to remove those in 8d15b19, which
resulted in some roads (and presumably other features) not being styled
properly.