[deploy] fixup docker publish script.
For some reason `earthly --push +images` *appears* to succeed, but the
dev tag isn't updated. Interestingly, when I do a docker --push (e.g.
via this bin/publish-images script), it shows that the layers already
exist on the remote, but upon completion the tag is updated.

Maybe there's an error in tagging?