[build] update to latest stable OTP
I swear I was seeing an error when trying to start the server with
assets from this tagged OTP build, but now I'm seeing the reverse:
The same error but when using this pinned build.

So I don't know if I had just confused myself previously or what, but
I'm happy to be on a stable tagged build in any case.