[www] Revert SearchBox from Select back to Input widget
A few commits back, as part of working on the Search page, I replaced the SearchBox innards, which used to be based on a QInput instead with a QSelect widget.

There were some nice advantages to this out of the box (more or less):
- better layout
- keyboard navigation (up/down/enter-to-select)
- activity indicator

However there were also a few issues. Ultimately, I feel like adding the
above niceties to our own widget based on QInput will be a better way
forward rather than trying to hack around the relatively prescriptive
format of the QSelect.

I don't think QSelect is necessarily "broken" or anything, it's just
that we have a combination of several rather idiosyncratic requirements,

- autocomplete text to search for models (Place), but *leave* the user's raw
  input text untouched. Currently it's cleared on blur.
- refining continuous searches (I had a work around to keep the menu
  from flickering, but it was a hack)
- the debounce/ignore stale request logic required grabbing the raw
  input of the text (I had a work around for this too, but it was a
  hack).