[www] FIX: redundant autocomplete overlays subsequent search results
This was affecting *subsequent* searches (after you were already on the
/search/<something> page.

After typing "coffee<Enter>" you'd see the search results updated as
expected, but then you'd also see the autocomplete become visible,
overlaying the search results. It was redundant and "in the way".

This doesn't affect the initial search because you're taken to a new
page — the old page has it's own SearchBox, so it's unmounted and does
not display these (redundant) autocomplete results.