[frontend] avoid confusing error when serving stale assets
For our SPA to work, we route all requests for which a file can't be
found, to index.html - assuming it's a page request that the js app
router will handle.

But when loading a stale index.html, we'll request assets, which include
a stale cache-buster (like style.ab123.css). These should 404, but
instead are hitting the fallback rule and are served the content of
index.html

So we now have a special rule for css/js assets, and also mark them for
aggressive caching since they are immutable.