Clip OTP input from same source as map tiles
This is helpful for the planet build to be sure we're using the same map
data as input to the transit data as we use for tile building and
everything else.

Before this change, we'd use the downloaded planet.osm.pbf for tiles,
etc. but then use bbike extracts for the transit zones. In theory they
could be out of sync.

Other incidental changes:

- extract bbox logic into target
- apt install before bbox, there's no reason to constantly invalidate these layers