[tileserver] water above sand, waterways above water
See https://maps.earth/place/-124.2153059153271,47.24507569180591

The Moclips River dumps into the Pacific Ocean, but with the previous
rendering it appears like there is a bunch of sand between where the
river ends and the ocean begins.

The new layer ordering seems to be more consistent with how the map is
expected to be rendered.

Some related discussion:

https://help.openstreetmap.org/questions/49790/coastline-at-highwater-vs-lowwater

> It has been a long-held practice in OSM (2008 or earlier) to map
coastlines at the Mean High Water Spring as stated here: