[transit] Infer the OTP bounding box from the input GTFS
We need a big enough area that our gtfs routes aren't cropped, but since
OTP requires memory proportional to the input osm.pbf, we still want to
be able to crop out a minimal the area to pass into OTP.

Previously we were specifying the bbox value via an "area" parameter,
that corresponded to a set of hardcoded bboxes; however, this could be
problematic if the transit feeds wandered outside the bounds of that
area. So now we compute it from the actual data in the gtfs data.

Also:

- version control the demonstration gtfs_feeds
- switch the planet demo's seattle transit zone to use Sound
  Transit's consolidated Puget Sound GTFS and add Island Transit