[otp-build] Breaking: Update OTP to fix cross timezone transit routing
In my case (the "builds/planet" example), the Whatcom County GTFS
agency.txt lists the timezone as "America/Tijuana", whereas others in
the same graph use either "America/Los_Angeles" or do not specify.

When Whatcom County is in a stand-alone node (with no other feeds),
routable trips are found, and everything works as expected, but when
incorporating the Whatcom GTFS feed alongside the others, transitmux was
not able to find an OTP for the trip, implying that there were no routes
covering the area.

Updating to the latest OTP and rebuilding the transit graph fixes the
issue. Not this is not an officially tagged release - we're just pinning
to a nightly build.

Because this upgrades the OTP serialization version, this is a breaking
change. Existing deployments will need to rebuild their transit graph
and their containers.