[www] Expose OSM's opening hours
Note that the current "status" implementation ("it's currently closed")
assumes you're looking at places only within your own timezone, which is
obviously flawed.

I think addressing that will be non-trivial. Maybe we could plumb the
timezone of each place through pelias?

As it is, I believe it's the common case, and thus useful enough to
include, even with this limitation.