[ui] mark transfers / legs
Otherwise it can be hard to see right where you switch buses or start
walking.

The former is a bigger problem than the latter, but I think it's worth
showing for pedestrian leg start/ends as well for design consistency
with the transit timeline "nodes" (in the left panel/bottom card).