[basemap] show states when zoomed out farther
Currently state boundaries are only visible when zoomed way in.

For small countries with relatively many states this could get a bit
cluttered when zoomed out. I tried to balance the design trade-offs
between large countries and small countries.

It'd be nice if we knew which country (admin_2) that a given "state"
boundary (admin_level_3/4) belonged to, but this isn't included in the
opentiles schema.

I might look at modifying planetiler for this purpose at some point, but
for right now it seems like more work than I'm prepared to commit.