WIP: Valhalla integration into transitmux
Currently this returns a "one of" response, but it did allow me to plumb
through all the configuration we need to talk between OTP and Valhalla.

And I did unify the "query" api - though it is limited to the subset
of features we currently use.

The big remaining thing will be unifying the output so that a transitmux
client can render trips from opentripplanner or valhalla with more or
less the same code.