[www] Fix valhalla routing instructions
Clicking on a route step and non-transit routing was not rendering.

I broke this while working on the travelmux API unification. I
unwittingly committed some code to the wrong file that made valhalla api
responses camelcase.

OTP is camelCase and Valhalla is snake_case. I decided to adopt
camelCase for the unified responses, but my intention was always to
leave the original _valhalla and _otp responses untouched so that legacy
clients could easily migrate.