adapt to OTP 2.5 config
I saw this in the logs:

    16:54:50.419 WARN [main]  (NodeAdapter.java:170) Unexpected config parameter: 'updaters.[0].frequencySec:60' in 'router-config.json'
    16:54:50.419 WARN [main]  (NodeAdapter.java:170) Unexpected config parameter: 'updaters.[1].frequencySec:300' in 'router-config.json'

According to git, I updated the containers to 2.5.0 a year ago - so I'm
not sure if some default value was being assumed or what.

I noticed this because the puget sount OTP containers was failing to
start. I haven't yet tracked downt he cause, and these were just
warnings, so I'm assuming they're not responsible for the startup
failure.