Integration tests
A full build/test involves a lot of waiting around before some more
manual testing.

We're already building all the assets in CI, but we didn't have any
end-to-end tests with building and standing up services using that data.

I'm hoping this will serve as a foundation to speed up accepting of
outside contribution in particular.