Simplify env
It was getting too confusing mixing the docker .env, the dagger .env,
and the env used for scripts. Now those are separate. The env used for
scripts lives at .bin-env and the env used for docker and dagger lives
in builds/<area>/.env

This means more commands require an explicit build dir, which is maybe
annoying, but it seems more sane.