Simplify env
It was getting too confusing mixing the docker .env, the dagger .env,
and the env used for scripts. Now those are separate. The env used for
scripts lives at .bin-env and the env used for docker and dagger lives
in builds/<area>/.env

This means more commands require an explicit build-config dir, which is
maybe annoying, but it seems more sane.

Relatedly: Made the build-config dir match the case of HEADWAY_AREA
(sane filesystems are case sensitive)

Also: bin/link-transit now works locally